my rationale for breaking the world: HederaService was creating a client and functions that used mirrorNodeUrl, i don't think we were modelling objects so much as creating namespace so split those back into static utility functions + added mirrorNodeUrl as a parameter
