+++ packages/hedera-wallet-snap/packages/snap/src/rpc/transactions/getTransactions.ts
@@ -19,7 +19,7 @@
  */
 
 import { providerErrors } from '@metamask/rpc-errors';
-import { HederaServiceImpl } from '../../services/impl/hedera/service/HederaServiceImpl';
+import { HederaServiceImpl } from '../../client/impl/hedera/service/HederaServiceImpl';
 import { MirrorTransactionInfo } from '../../types/hedera';
 import { GetTransactionsRequestParams } from '../../types/params';
 import { WalletSnapParams } from '../../types/state';
